added user settings model
